backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
prometheus-api-client==0.5.3
kubernetes==28.1.0
aiohttp==3.9.1
python-dotenv==1.0.0
pyyaml==6.0.1
pydantic==2.5.2
